pyapi/biovenv/requirements.txt
biopython==1.81
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Werkzeug==2.2.3
